backend/pyproject.toml
[tool.poetry]
name = "backendproject"
version = "0.1.0"
description = ""
authors = ["AlexanderG1999 <95202925+AlexanderG1999@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "            3.7.2"
django = "             4.1.13"
django-cors-headers = "4.3.1"
djangorestframework = "3.14.0"
djongo = "             1.3.6"
dnspython = "          2.4.2"
pip = "                22.2.1"
pymongo = "            3.12.0"
pytz = "               2023.3.post1"
setuptools = "         63.2.0"
sqlparse = "           0.2.4"
typing-extensions = "  4.8.0"
tzdata = "             2023.3"
python-decouple = "^3.8"
pyjwt = "^2.8.0"
pillow = "^10.2.0"
boto3 = "^1.34.22"
django-storages = "^1.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
